chore: moved category module
