+++ src/modules/shop/shop-category/dto/create-shop-category.dto.ts
@@ -1,6 +1,6 @@
 import { Shop } from 'src/modules/shop/shop/entities/shop.entity';
 import { IsNotEmpty } from 'class-validator';
-import { Category } from '../../category/entities/category.entity';
+import { Category } from '../../../category/entities/category.entity';
 
 export class CreateShopCategoryDto {
   @IsNotEmpty()
