+++ src/modules/shop/shop-category/entities/shop-category.entity.ts
@@ -7,7 +7,7 @@ import {
 } from 'typeorm';
 import { Shop } from 'src/modules/shop/shop/entities/shop.entity';
 import { Product } from 'src/modules/product/entities/product.entity';
-import { Category } from '../../category/entities/category.entity';
+import { Category } from '../../../category/entities/category.entity';
 
 @Entity()
 export class ShopCategory {
