+++ src/modules/shop/shop/shop.module.ts
@@ -9,8 +9,8 @@ import { ShopCategoryService } from '../shop-category/shop-category.service';
 import { Shop } from './entities/shop.entity';
 import { WorkingHour } from '../working_hours/entities/working_hour.entity';
 import { ShopCategory } from '../shop-category/entities/shop-category.entity';
-import { CategoryService } from '../category/category.service';
-import { Category } from '../category/entities/category.entity';
+import { CategoryService } from '../../category/category.service';
+import { Category } from '../../category/entities/category.entity';
 
 @Module({
   imports: [
